Reorganize bootstrap modules structure
