+++ src/boot/bootstrap/bootstrap-components.spec.ts
@@ -6,13 +6,15 @@ import { Component } from '../../component';
 import { FeatureContext, FeatureDef, LoadedFeature } from '../../feature';
 import { MethodSpy } from '../../spec/mocks';
 import { BootstrapContext } from '../bootstrap-context';
-import { ComponentRegistry } from '../definition/component-registry.impl';
-import { ComponentValueRegistry } from '../definition/component-value-registry.impl';
-import { DefinitionValueRegistry } from '../definition/definition-value-registry.impl';
-import { ElementBuilder } from '../definition/element-builder.impl';
 import { DefaultNamespaceAliaser } from '../globals';
+import {
+  BootstrapValueRegistry,
+  ComponentRegistry,
+  ComponentValueRegistry,
+  DefinitionValueRegistry,
+  ElementBuilder,
+} from '../impl';
 import { bootstrapComponents } from './bootstrap-components';
-import { BootstrapValueRegistry } from './bootstrap-value-registry.impl';
 import Mock = jest.Mock;
 
 describe('boot', () => {
