+++ src/boot/impl/component-factory.symbol.impl.ts
