+++ src/boot/impl/component-registry.impl.ts
