+++ src/boot/impl/index.ts
@@ -1,0 +1,5 @@
+export * from './bootstrap-value-registry.impl';
+export * from './component-registry.impl';
+export * from './component-value-registry.impl';
+export * from './definition-value-registry.impl';
+export * from './element-builder.impl';
