+++ src/feature/loader/feature-load.spec.ts
@@ -2,10 +2,12 @@ import { valueProvider } from 'call-thru';
 import { ContextValues, SingleContextKey, SingleContextUpKey } from 'context-values';
 import { AfterEvent, eventInterest, EventInterest } from 'fun-events';
 import { BootstrapContext } from '../../boot';
-import { BootstrapValueRegistry } from '../../boot/bootstrap/bootstrap-value-registry.impl';
-import { ComponentRegistry } from '../../boot/definition/component-registry.impl';
-import { ComponentValueRegistry } from '../../boot/definition/component-value-registry.impl';
-import { DefinitionValueRegistry } from '../../boot/definition/definition-value-registry.impl';
+import {
+  BootstrapValueRegistry,
+  ComponentRegistry,
+  ComponentValueRegistry,
+  DefinitionValueRegistry,
+} from '../../boot/impl';
 import { Class } from '../../common';
 import { ComponentDef } from '../../component';
 import { FeatureContext } from '../feature-context';
