+++ src/feature/loader/feature-loader.impl.ts
@@ -12,10 +12,12 @@ import {
   trackValue,
 } from 'fun-events';
 import { BootstrapContext } from '../../boot';
-import { BootstrapValueRegistry } from '../../boot/bootstrap/bootstrap-value-registry.impl';
-import { ComponentRegistry } from '../../boot/definition/component-registry.impl';
-import { ComponentValueRegistry } from '../../boot/definition/component-value-registry.impl';
-import { DefinitionValueRegistry } from '../../boot/definition/definition-value-registry.impl';
+import {
+  BootstrapValueRegistry,
+  ComponentRegistry,
+  ComponentValueRegistry,
+  DefinitionValueRegistry,
+} from '../../boot/impl';
 import { ArraySet, Class } from '../../common';
 import { ComponentContext } from '../../component';
 import { ComponentClass, DefinitionContext } from '../../component/definition';
